+++ monopyly/analytics/routes.py
@@ -7,6 +7,7 @@ from flask import g, render_template, request
 
 from ..auth.tools import login_required
 from ..common.transactions import TransactionTagHandler
+from .actions import get_tag_statistics_chart_data
 from .blueprint import bp
 
 
@@ -49,3 +50,30 @@ def delete_tag():
     # Remove the tag from the database
     TransactionTagHandler.delete_entry(tag.id)
     return ""
+
+
+@bp.route("/tag_statistics")
+@login_required
+def show_tag_statistics():
+    hierarchy = TransactionTagHandler.get_hierarchy()
+    return render_template(
+        "analytics/tag_statistics/tag_statistics_page.html",
+        tags_hierarchy=hierarchy,
+        chart_data=get_tag_statistics_chart_data(hierarchy.keys()),
+    )
+
+
+@bp.route("/_update_tag_statistics_chart", methods=("POST",))
+@login_required
+def update_tag_statistics_chart():
+    # Get the field from the AJAX request
+    tag_id = int(request.get_json())
+    tags = (
+        [TransactionTagHandler.get_entry(tag_id)]
+        if tag_id
+        else TransactionTagHandler.get_hierarchy()
+    )
+    return render_template(
+        "analytics/tag_statistics/tag_statistics.html",
+        chart_data=get_tag_statistics_chart_data(tags),
+    )